requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
PyYAML>=6.0

# Deep learning utilities
tensorboard>=2.12.0
tqdm>=4.65.0
tensorflow>=2.12.0

# Computer vision
opencv-python>=4.7.0
Pillow>=9.4.0

# Data handling
pandas>=1.5.0
scipy>=1.10.0

# For CLIP
ftfy>=6.1.1
regex>=2023.3.23

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
pyyaml>=6.0
pathlib>=1.0.1

